Refactor JmxMetricsWatcher and split its responsibilities
