+++ bootstrap/src/main/java/com/splunk/opentelemetry/javaagent/bootstrap/jmx/JmxListener.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright Splunk Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.splunk.opentelemetry.javaagent.bootstrap.jmx;
+
+import javax.management.JMException;
+import javax.management.MBeanServer;
+import javax.management.ObjectName;
+
+public interface JmxListener {
+  void onRegister(MBeanServer mBeanServer, ObjectName objectName) throws JMException;
+
+  void onUnregister(MBeanServer mBeanServer, ObjectName objectName) throws JMException;
+}
